+++ src/engine/component_storage.cc
@@ -1,0 +1,8 @@
+#include "engine/component_storage.h"
+
+
+namespace petri::engine {
+
+
+
+} // namespace petri::engine