+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## [0.1.0e] – Initial Engine bootstrap
 - Time counter with managing `DeltaSeconds`, `ElapsedSeconds`, `TimeScale`, `IsPaused`
 - Engine class bootstrap with Update() and managing Time object
+- World class bootstrap with Init() and Tick() placeholders
+- Entity struct and EntityStorage placeholder that stores dense array of Entities and helpers needed for mapping dense to sparse
+- Minimal ComponentStorage and basic components (Position, Velocity, Size, Energy) with traits to compile-time resolving of ComponentStorage::Get<ComponentType>()
+- Systems API placeholder with dump MakeMovement() system
+- Time estimation changed to float
 
 ## [0.1.0] – Initial App bootstrap